refactor: useRefresh 훅으로 정리 #37
